requirements.txt
flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.1
torch==1.10.0
torchvision==0.11.1
numpy==1.21.4
Pillow==8.4.0
matplotlib==3.4.3
scikit-learn==1.0.1
pandas==1.3.4
seaborn==0.11.2
tqdm==4.62.3
opencv-python==4.5.4.58
pytest==6.2.5
pytest-flask==1.2.0
datasets==2.10.1